Add MIT license
This commit adds the MIT license to the project, granting users the freedom to use, modify, distribute, and sell the software. It provides clear permission to include the copyright notice in all copies or substantial portions of the software. The addition of this license increases the project's openness and encourages collaboration and reuse.