Refactor shipping cost calculations to use integer ID values
The commit refactors the shipping cost calculations in the `get_shipping_cost` and `calculate_shipping_cost` functions to use integer ID values for the origin and destination instead of strings. This change improves efficiency by using more compact data types and aligning with the ID-based structure of the `get_subdistricts` function.